Moved interfaces to new module
